requirements.txt
pydicom==2.4.4
pandas==2.2.2
sqlalchemy==2.0.32
pyodbc==5.1.0
python-dotenv==1.0.1
tqdm==4.66.4
